+++ app/src/androidTest/java/nextstep/github/ui/github/GithubScreenTest.kt
@@ -1,4 +1,4 @@
-package nextstep.github
+package nextstep.github.ui.github
 
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.SnackbarResult
@@ -12,8 +12,6 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.test.advanceUntilIdle
 import kotlinx.coroutines.test.runTest
 import nextstep.github.model.Repository
-import nextstep.github.ui.github.GithubScreen
-import nextstep.github.ui.github.GithubUiState
 import org.junit.Rule
 import org.junit.Test
