+++ app/src/androidTest/java/nextstep/github/ui/github/component/GithubLoadingTest.kt
@@ -1,9 +1,8 @@
-package nextstep.github.component
+package nextstep.github.ui.github.component
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithTag
-import nextstep.github.ui.github.component.GithubLoading
 import org.junit.Rule
 import org.junit.Test
